requirements.txt
biopython==1.70
numpy==1.14.0
wget==3.2